operator/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitarray==3.4.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
cytoolz==1.0.1
dnspython==2.7.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.115.12
frozenlist==1.7.0
h11==0.16.0
hexbytes==1.3.1
idna==3.10
motor==3.7.1
multidict==6.4.4
parsimonious==0.10.0
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.1
PyNaCl==1.5.0
python-dotenv==1.1.0
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
sniffio==1.3.1
sparse-merkle-tree==0.3.0
starlette==0.46.2
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
web3==7.12.0
websockets==15.0.1
yarl==1.20.1
